integracao da api da pagina resultado e map
