+++ src/App.js
@@ -3,8 +3,10 @@ import Router from "./Routes/Router"
 import theme from './constants/theme'
 import { ThemeProvider } from '@material-ui/core/styles'
 
+
 const App = () => {
 
+  
   return (
 
     <ThemeProvider theme={theme}>
