+++ src/Pages/Resultado/style.js
@@ -29,6 +29,13 @@ div{
     flex-direction:row;
 }
 
+h4{
+    margin-right:40px;
+    margin-bottom:0.5px;
+    margin-top:8px;
+    color:red
+    
+}
 
 p{
     margin-right:40px;
